Consolidate web routes
