+++ apps/web/core/controllers.rb
@@ -1,13 +1,7 @@
 # apps/web/core/controllers.rb
 
 require_relative 'controllers/page'
-require_relative 'controllers/account'
 require_relative 'controllers/welcome'
 require_relative 'controllers/authentication'
 require_relative 'controllers/registration'
-require_relative 'controllers/session'
-require_relative 'controllers/password_reset'
-require_relative 'controllers/image'
-require_relative 'controllers/robots'
-require_relative 'controllers/data_export'
 require_relative 'controllers/data_incoming'
