+++ apps/web/core/controllers/authentication.rb
@@ -27,8 +27,8 @@ module Core
 
         logic = V2::Logic::Authentication::AuthenticateSession.new(strategy_result, req.params, locale)
 
-        if session.authenticated?
-          session.set_info_message 'You are already logged in.'
+        if authenticated?
+          session['info_message'] = 'You are already logged in.'
           res.redirect '/'
         elsif req.post?
           logic.raise_concerns
@@ -36,7 +36,7 @@ module Core
           sess = logic.sess
           cust_after = logic.cust
 
-          res.send_secure_cookie :sess, session.sessid, session.default_expiration
+          # Session cookie handled by Rack::Session middleware
 
           if cust_after.role?(:colonel)
             res.redirect '/colonel/'
@@ -44,6 +44,26 @@ module Core
             res.redirect '/'
           end
         end
+      end
+
+      def logout
+        res.no_cache!
+
+        strategy_result = Otto::Security::Authentication::StrategyResult.new(
+          session: session,
+          user: cust,
+          auth_method: 'session',
+          metadata: {
+            ip: req.client_ipaddress,
+            user_agent: req.user_agent
+          }
+        )
+
+        logic = V2::Logic::Authentication::DestroySession.new(strategy_result, req.params, locale)
+        logic.raise_concerns
+        logic.process
+
+        res.redirect res.app_path('/')
       end
 
       private
