+++ apps/web/core/controllers/registration.rb
@@ -31,6 +31,30 @@ module Core
         res.redirect '/'
       end
 
+      def request_reset
+        strategy_result = Otto::Security::Authentication::StrategyResult.new(
+          session: session,
+          user: cust,
+          auth_method: 'session',
+          metadata: {
+            ip: req.client_ipaddress,
+            user_agent: req.user_agent
+          }
+        )
+
+        if req.params[:key]
+          logic = V2::Logic::Authentication::ResetPassword.new(strategy_result, req.params, locale)
+          logic.raise_concerns
+          logic.process
+          res.redirect '/signin'
+        else
+          logic = V2::Logic::Authentication::ResetPasswordRequest.new(strategy_result, req.params, locale)
+          logic.raise_concerns
+          logic.process
+          res.redirect '/'
+        end
+      end
+
       private
 
       def _auth_settings
