--- apps/web/core/controllers/session.rb
@@ -1,31 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'base'
-
-module Core
-  module Controllers
-    class Session
-      include Controllers::Base
-
-      def logout
-        res.no_cache!
-
-        strategy_result = Otto::Security::Authentication::StrategyResult.new(
-          session: session,
-          user: cust,
-          auth_method: 'session',
-          metadata: {
-            ip: req.client_ipaddress,
-            user_agent: req.user_agent
-          }
-        )
-
-        logic = V2::Logic::Authentication::DestroySession.new(strategy_result, req.params, locale)
-        logic.raise_concerns
-        logic.process
-
-        res.redirect res.app_path('/')
-      end
-    end
-  end
-end
